feat: Implement AgentsGalleryModal with enhanced UI and localization
