+++ src/views/Brain/RouterAgentsTeam/AgentsGalleryModal.vue
@@ -1,0 +1,277 @@
+<template>
+  <UnnnicModalDialog
+    class="agents-gallery-modal"
+    :modelValue="agentsTeamStore.isAgentsGalleryOpen"
+    showCloseIcon
+    showActionsDivider
+    :title="$t('router.agents_team.gallery.title')"
+    size="lg"
+    @update:model-value="closeModal"
+  >
+    <section
+      class="agents-gallery"
+      data-testid="agents-gallery"
+    >
+      <SidebarMenu class="agents-gallery__sidebar">
+        <SideBarItem
+          v-for="tab in contentTabs"
+          :key="tab.page"
+          class="sidebar__item"
+          :text="$t(`router.agents_team.gallery.${tab.title}`)"
+          :active="tab.page === activeTab"
+          :iconFilled="tab.page === activeTab"
+          data-test="nav-agent-type"
+          @click="onTabChange(tab.page)"
+        />
+      </SidebarMenu>
+
+      <section
+        v-if="isMyAgentsEmpty && activeTab === 'my-agents'"
+        class="agents-gallery__empty-custom-agents"
+      >
+        <UnnnicIntelligenceText
+          tag="p"
+          family="secondary"
+          size="body-gt"
+          color="neutral-cloudy"
+        >
+          {{ $t('router.agents_team.gallery.custom_agents_description') }}
+        </UnnnicIntelligenceText>
+
+        <!-- This comment prevents from auto-capitalizing i18n-t to I18nT which would break the component -->
+        <!-- eslint-disable-next-line vue/component-name-in-template-casing -->
+        <i18n-t
+          class="empty-custom-agents__description"
+          keypath="router.agents_team.gallery.weni_cli_documentation_description"
+          tag="p"
+        >
+          <template #weni_cli_documentation>
+            <p
+              class="description__link"
+              @click="openCLI"
+            >
+              {{ $t('router.agents_team.gallery.weni_cli_documentation') }}
+            </p>
+          </template>
+        </i18n-t>
+      </section>
+
+      <section
+        v-else
+        class="agents-gallery__content"
+      >
+        <UnnnicInput
+          v-model="search[activeTab]"
+          iconLeft="search"
+          :placeholder="$t('router.agents_team.gallery.search_placeholder')"
+          data-testid="search-input"
+        />
+
+        <section
+          :class="[
+            'content__cards',
+            {
+              'content__cards--empty': isMyAgentsEmpty,
+            },
+          ]"
+          data-testid="agent-cards"
+        >
+          <template v-if="isLoadingAgents">
+            <AssignAgentCard
+              v-for="(_, index) in Array(3)"
+              :key="index"
+              loading
+              data-testid="agent-card-loading"
+            />
+          </template>
+
+          <template v-else>
+            <UnnnicIntelligenceText
+              v-if="isSearchEmpty"
+              color="neutral-clean"
+              family="secondary"
+              size="body-gt"
+              tag="p"
+            >
+              {{ $t('router.agents_team.gallery.no_agent_found') }}
+            </UnnnicIntelligenceText>
+
+            <template v-else>
+              <AssignAgentCard
+                v-for="agent in agentsData"
+                :key="agent.uuid"
+                :agent="agent"
+                data-testid="agent-card"
+              />
+            </template>
+          </template>
+        </section>
+      </section>
+    </section>
+  </UnnnicModalDialog>
+</template>
+
+<script setup>
+import { computed, onMounted, ref, watch } from 'vue';
+import { debounce } from 'lodash';
+
+import AssignAgentCard from '@/components/Brain/AgentsTeam/AssignAgentCard.vue';
+import SideBarItem from '@/components/Sidebar/SideBarItem.vue';
+import SidebarMenu from '@/components/Sidebar/SidebarMenu.vue';
+
+import { useAgentsTeamStore } from '@/store/AgentsTeam';
+
+const agentsTeamStore = useAgentsTeamStore();
+const officialAgents = agentsTeamStore.officialAgents;
+const myAgents = agentsTeamStore.myAgents;
+
+const contentTabs = ref([
+  { title: 'official_agents', page: 'official' },
+  { title: 'custom_agents', page: 'my-agents' },
+]);
+const activeTab = ref(contentTabs.value[0].page);
+const search = ref({
+  official: '',
+  'my-agents': '',
+});
+const isSearchEmpty = ref(false);
+
+const agentsData = computed(() =>
+  activeTab.value === 'official' ? officialAgents.data : myAgents.data,
+);
+const isLoadingAgents = computed(
+  () => officialAgents.status === 'loading' || myAgents.status === 'loading',
+);
+const isMyAgentsEmpty = computed(
+  () =>
+    activeTab.value === 'my-agents' &&
+    myAgents.data.length === 0 &&
+    !isLoadingAgents.value,
+);
+
+function updateSearchEmpty() {
+  const currentAgents = agentsData.value;
+  const currentSearch = search.value[activeTab.value];
+
+  isSearchEmpty.value =
+    currentSearch.trim() !== '' && currentAgents.length === 0;
+}
+
+function onTabChange(newTab) {
+  activeTab.value = newTab;
+
+  if (agentsTeamStore.myAgents.status === null && newTab === 'my-agents') {
+    agentsTeamStore.loadMyAgents();
+  }
+
+  updateSearchEmpty();
+}
+
+function closeModal() {
+  agentsTeamStore.isAgentsGalleryOpen = false;
+}
+
+function openCLI() {
+  window.open(agentsTeamStore.linkToCreateAgent, '_blank');
+}
+
+onMounted(() => {
+  agentsTeamStore.loadOfficialAgents();
+});
+
+const debouncedSearch = (callback) => debounce(callback, 300);
+watch(
+  () => search.value['my-agents'],
+  debouncedSearch(async (search) => {
+    await agentsTeamStore.loadMyAgents({ search });
+    updateSearchEmpty();
+  }),
+);
+watch(
+  () => search.value['official'],
+  debouncedSearch(async (search) => {
+    await agentsTeamStore.loadOfficialAgents({ search });
+    updateSearchEmpty();
+  }),
+);
+</script>
+
+<style lang="scss" scoped>
+.agents-gallery-modal {
+  :deep(.unnnic-modal-dialog__container) {
+    width: 75vw;
+    height: 85vh;
+  }
+
+  &__sidebar {
+    .sidebar__item {
+      padding: $unnnic-spacing-ant;
+    }
+  }
+
+  .agents-gallery {
+    height: 100%;
+
+    display: grid;
+    grid-template-columns: 170px 1fr;
+    gap: $unnnic-spacing-sm;
+
+    &__content {
+      display: flex;
+      flex-direction: column;
+      gap: $unnnic-spacing-sm;
+
+      .content__cards {
+        display: grid;
+        grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+        gap: $unnnic-spacing-sm;
+
+        &--empty {
+          grid-template-columns: repeat(1, 1fr);
+        }
+      }
+    }
+
+    &__empty-custom-agents {
+      width: 100%;
+      height: 100%;
+
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      flex-direction: column;
+
+      .empty-custom-agents__description {
+        display: flex;
+        align-items: center;
+        gap: $unnnic-spacing-nano;
+
+        font-size: $unnnic-font-size-body-gt;
+
+        .description__link {
+          position: relative;
+
+          font-size: $unnnic-font-size-body-gt;
+          color: $unnnic-color-neutral-cloudy;
+          font-weight: $unnnic-font-weight-bold;
+
+          cursor: pointer;
+
+          &::after {
+            content: '';
+
+            position: absolute;
+            bottom: 0;
+            left: 0;
+
+            width: 100%;
+            height: 1px;
+
+            background-color: $unnnic-color-neutral-cloudy;
+          }
+        }
+      }
+    }
+  }
+}
+</style>
